add memory optimizations, enable warnings in validation pipeline
- uses optimization techs borrowed from blissful-tuner (https://github.com/Sarania/blissful-tuner)

+++ src/modules/fp8_optimization_utils.py
@@ -87,9 +87,11 @@ def quantize_tensor_to_fp8(
     # Calculate log scales (only for non-zero elements)
     log_scales = torch.zeros_like(clamped_tensor)
     if nonzero_mask.any():
-        log_scales[nonzero_mask] = torch.floor(
-            torch.log2(abs_values[nonzero_mask]) + bias
-        ).detach()
+        log_scales[nonzero_mask] = (
+            torch.floor(torch.log2(abs_values[nonzero_mask]) + bias)
+            .detach()
+            .to(log_scales.dtype)
+        )
 
     # Limit log scales and calculate quantization factor
     log_scales = torch.clamp(log_scales, min=1.0)
